Update Domain 1- Cloud Computing Concepts and Architectures.md
Line 92 explanation of SPI should be  Services, Platform and Infrastructure instead of Application infrastructure and Platform.

Line 187 Meta structure and management planes is good, but there is an element of abstraction woven in there which provides isolation along with Virtualization. Also there is the benefit of abstraction from hardware failure at the application layer which provides infinite availability from an end user perspective along with elasticity. That is another key difference between Virtual hosting and Cloud environments- the abstraction layer.

On the whole the concept of tenant isolation for IaaS, PaaS and SaaS should be mentioned somewhere in this chapter as that is critical from a security standpoint and how that is achieved by different layers and different patterns for the same. 

In reality most SaaS solutions are not REAL Cloud Services, they are virtually hosted applications provided as services, that includes even services like Office 365..which is a hosted service and not a Cloud service with actual elasticity...at the app /services layer.

Another concept which didn't come thru is Cloud Broker services, and they may or may not be Cloudy.

There was some mention of Provider Chaining but the concept didn't come thru as that brings up concerns for Security and Integrations end to end for an Enterprise.

For Shared security and Operations model may be there can be a Matrix which shows the divide of responsibilities or at least some examples of the same for IaaS, PaaS and SaaS ( SaaS is the easiest obviously)

For hybrid Cloud enterprises there are some key challenges which should be highlighted (Maybe) or referenced somehow to another paper. There can be multiple permutations of hybrid Cloud Enterprises as well.

Another concept that can be referenced in this paper is: Inter Cloud and what that means for a Hybrid Cloud Enterprise Architecture.
Containerization even though may be addressed somewhere else, but a mention of the technology in the architectural concept may be nice to have in this section (but either way will work) as long as it is referenced here and details can be provided else where in the overall Guidance.